Create new controller for authentication and account linking
