+++ kifi-backend/modules/shoebox/app/com/keepit/controllers/ext/ExtAuthController.scala
@@ -10,8 +10,6 @@ import com.keepit.common.net._
 import com.keepit.model._
 
 import play.api.libs.json.Json
-import play.api.mvc.Action
-import securesocial.core._
 
 @Singleton
 class ExtAuthController @Inject() (
@@ -25,29 +23,6 @@ class ExtAuthController @Inject() (
   userExperimentRepo: UserExperimentRepo,
   kifiInstallationCookie: KifiInstallationCookie)
   extends BrowserExtensionController(actionAuthenticator) with ShoeboxServiceController {
-
-  // Some of this is copied from ProviderController in SecureSocial
-  // We might be able to do this better but I'm just trying to get it working for now
-  def mobileAuth(providerName: String) = Action(parse.json) { implicit request =>
-    // e.g. { "accessToken": "..." }
-    val oauth2Info = Json.fromJson(request.body)(Json.reads[OAuth2Info]).asOpt
-    val provider = Registry.providers.get(providerName).get
-    val authMethod = provider.authMethod
-    val filledUser = provider.fillProfile(
-      SocialUser(UserId("", providerName), "", "", "", None, None, authMethod, oAuth2Info = oauth2Info))
-    UserService.find(filledUser.id) map { user =>
-      val withSession = Events.fire(new LoginEvent(user)).getOrElse(session)
-      Authenticator.create(user) match {
-        case Right(authenticator) =>
-          Ok(Json.obj("sessionId" -> authenticator.id)).withSession(
-              withSession - SecureSocial.OriginalUrlKey - IdentityProvider.SessionId - OAuth1Provider.CacheKey)
-            .withCookies(authenticator.toCookie)
-        case Left(error) => throw error
-      }
-    } getOrElse {
-      NotFound(Json.obj("error" -> "user not found"))
-    }
-  }
 
   def start = AuthenticatedJsonToJsonAction { implicit request =>
     val userId = request.userId
